Uber India Demand-Supply Gap Analysis
This upload contains R-code script file written to analyze Uber-India's dataset to identify underlying patterns resulting in driver initiated cancellation and non-availability of taxis particularly for airport trips leading to loss of revenue for Uber.
The attached pdf file presents the results of the aforementioned analysis and suggested remedial measures to mitigate the problem. 

Please Note: Any future or current students of the PGDDS program are advised to refrain from copying segments of this code in their submissions as each submission is passed through a plagiarism checker and it will be identified.