Risk Analysis of Loan Application for Credit Loss
This upload contains the R-script file for the Gramener Case Study on Risk Analytics. It is aimed at applying structured Exploratory Data Analysis to understanding the concept of Risk Analytics and detect the factors driving Credit Loss. The insights generated are presented in the form of a condensed presentation attached as a pdf File.