Human Resource Analytics
This upload consists of the R-script file of a Logistic Regression Classifier used to Forecast if an employee is likely to Attrition from an organization. The model analyzes a number of parameters including employee details, employment history, educational details, job appraisal feedback and employee satisfaction feedback. The final model was evaluated for its deterministic performance. The model performance and results of the analysis are neatly presented in the attached pdf file. The presentation also includes critical suggestions to help curb the Employee Attrition Rate.