Digit Recognizer-MNIST Dataset with SVM-Classifer
The objective of the Digit Recognizer is to use the normalized 28x28 pixel information to predict the handwritten digit label. The handwritten digits vary from 0-9. Each Digit has 784 attributes corresponding to pixel density values between 0-255. I have implemented variance analysis strategy to drastically reduce the number of attributes to 250 for each digit label. I have built an SVM classifier with an RBF kernel and tuned the Hyperparameters through 3-fold cross-validation to achieve a classification accuracy of 97.63%.