Demand and Sales Forecasting using Time Series
This file contains the R-script file and the presentation pdf file for a forecasting model we built using Time Series Algorithm. We performed a market segment analysis to identify the top 3 consistently profitable region-consumer segment. For the aforementioned market segments, we built a demand and sales forecasting model to assist the operations and logistics teams in planning the sourcing and distribution strategy. The model was built using the classical decomposition method for Time Series Algorithm. We achieved a MAPE of less than 18% for all models.